HIVE-20388 : Move common classes out of metastore-server (Alexander Kolbasov reviewed by Vihang Karajgaonkar)
